feat: Merge the mmlu scripts into one
